+++ CHANGELOG.md
@@ -10,11 +10,12 @@ and this project adheres to
 
 ## Added
 
-- Allow to filter contracts by signature state, 
+- Add API endpoints for other services to fetch  data on course run
+- Allow to filter contracts by signature state,
   product, course and organization and id
 - Add bulk download of signed contracts to generate ZIP archive with command
 - Add read-only api admin endpoint to list/retrieve orders
-- Add a management command to synchronize course run or product 
+- Add a management command to synchronize course run or product
   on a remote catalog
 - Install and configure celery with redis
 - Add CachedModelSerializer